requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_select=2.1.0=gpu
absl-py=0.6.1=py36_0
astor=0.7.1=py36_0
backcall=0.1.0=py36_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1
ca-certificates=2018.03.07=0
certifi=2018.10.15=py36_0
cudatoolkit=9.2=0
cudnn=7.2.1=cuda9.2_0
cupti=9.2.148=0
decorator=4.3.0=py36_0
freetype=2.9.1=h8a8886c_1
gast=0.2.0=py36_0
grpcio=1.16.1=py36hf8bcb03_1
h5py=2.8.0=py36h989c5e5_3
hdf5=1.10.2=hba1933b_1
intel-openmp=2019.0=118
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.1.1=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
jedi=0.13.1=py36_0
jpeg=9b=habf39ab_1
jupyter_client=5.2.3=py36_0
jupyter_core=4.4.0=py36_0
keras-applications=1.0.6=py36_0
keras-base=2.2.4=py36_0
keras-gpu=2.2.4=0
keras-preprocessing=1.0.5=py36_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=h4deb6c0_3
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.35=hbc83047_0
libprotobuf=3.6.1=hd408876_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
markdown=3.0.1=py36_0
mkl=2019.0=118
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
ncurses=6.1=hf484d3e_0
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
olefile=0.46=py36_0
openssl=1.1.1=h7b6447c_0
parso=0.3.1=py36_0
pexpect=4.6.0=py36_0
pickleshare=0.7.5=py36_0
pillow=5.3.0=py36h34e0f95_0
pip=18.1=py36_0
prompt_toolkit=2.0.7=py36_0
protobuf=3.6.1=py36he6710b0_0
ptyprocess=0.6.0=py36_0
pygments=2.2.0=py36_0
python=3.6.7=h0371630_0
python-dateutil=2.7.5=py36_0
pyyaml=3.13=py36h14c3975_0
pyzmq=17.1.2=py36h14c3975_0
readline=7.0=h7b6447c_5
scikit-learn=0.20.0=py36h4989274_1
scipy=1.1.0=py36hfa4b5c9_1
setuptools=40.6.2=py36_0
six=1.11.0=py36_1
sqlite=3.25.3=h7b6447c_0
tensorboard=1.12.0=py36hf484d3e_0
tensorflow=1.12.0=gpu_py36he74679b_0
tensorflow-base=1.12.0=gpu_py36had579c0_0
tensorflow-gpu=1.12.0=h0d30ee6_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
tornado=5.1.1=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
werkzeug=0.14.1=py36_0
wheel=0.32.2=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=h96e3832_1
zeromq=4.2.5=hf484d3e_1
zlib=1.2.11=hfbfcf68_1
